Refactor script

+++ setup.sh
@@ -7,8 +7,12 @@ if [ -z "$git_email" ]; then
   exit 1
 fi
 
+cmd_missing() {
+ ! command -v "$1" &> /dev/null
+}
+
 setup_git() {
-  if ! command -v git; then
+  if cmd_missing git; then
     sudo apt install git -y &
   fi
   git config --global core.editor "vim"
@@ -21,7 +25,7 @@ setup_git() {
 }
 
 setup_fish() {
-  if ! command -v fish; then
+  if cmd_missing fish; then
     sudo apt-add-repository ppa:fish-shell/release-3 -y
     sudo apt update
     sudo apt install fish -y
@@ -29,7 +33,7 @@ setup_fish() {
   sudo chsh -s "$(which fish)" "$USER"
 
   # TODO: Replace with tide?
-  if ! command -v omf; then
+  if cmd_missing omf; then
     curl https://raw.githubusercontent.com/oh-my-fish/oh-my-fish/master/bin/install >install &&
       chmod +x install
     ./install --noninteractive
@@ -42,7 +46,7 @@ setup_fish() {
 }
 
 setup_nvm() {
-  if ! command -v nvm; then
+  if cmd_missing nvm; then
     fish -c "fisher install jorgebucaran/nvm.fish"
     fish -c "nvm install latest"
     fish -c "set --universal nvm_default_version latest"
@@ -50,7 +54,7 @@ setup_nvm() {
 }
 
 setup_homebrew() {
-  if ! command -v brew; then
+  if cmd_missing brew; then
     echo -ne '\n' | /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
     fish -c "fish_add_path /home/linuxbrew/.linuxbrew/bin"
   fi
@@ -68,7 +72,7 @@ setup_dependencies() {
 }
 
 setup_chrome() {
-  if ! command -v google-chrome; then
+  if cmd_missing google-chrome; then
     sudo apt install fonts-liberation &&
     wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb &&
       sudo apt install -y ./google-chrome-stable_current_amd64.deb &&
@@ -82,12 +86,12 @@ setup_dotfiles() {
 }
 
 setup_httpie(){
-    if ! command -v http; then
-  curl -SsL https://packages.httpie.io/deb/KEY.gpg | sudo gpg --dearmor -o /usr/share/keyrings/httpie.gpg
-sudo echo "deb [arch=amd64 signed-by=/usr/share/keyrings/httpie.gpg] https://packages.httpie.io/deb ./" > /etc/apt/sources.list.d/httpie.list
-sudo apt update
-sudo apt install httpie
-    fi
+  if cmd_missing http; then
+    curl -SsL https://packages.httpie.io/deb/KEY.gpg | sudo gpg --dearmor -o /usr/share/keyrings/httpie.gpg
+    sudo echo "deb [arch=amd64 signed-by=/usr/share/keyrings/httpie.gpg] https://packages.httpie.io/deb ./" > /etc/apt/sources.list.d/httpie.list
+    sudo apt update
+    sudo apt install httpie
+  fi
 }
 
 
@@ -100,44 +104,46 @@ setup_tools() {
   sudo cp bin/mousewheel.sh /usr/bin/scroll
 
   setup_httpie
-  # The Fuck
-  if ! command -v fuck; then
+
+  if cmd_missing fuck; then
     pipx install thefuck && pipx ensurepath && 
 	    sudo mv ~/.local/bin/thefuck /usr/bin/thefuck
 	    sudo mv ~/.local/bin/fuck /usr/bin/fuck
   fi
   # TODO: Ultrawide
-  # bun
-  curl -fsSL https://bun.sh/install | bash
+
+  if cmd_missing bun; then
+    curl -fsSL https://bun.sh/install | bash
+  fi
 }
 
 setup_docker() {
-  if ! command -v docker; then
-  # Add Docker's official GPG key:
-sudo apt-get update
-sudo apt-get install ca-certificates curl
-sudo install -m 0755 -d /etc/apt/keyrings
-sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
-sudo chmod a+r /etc/apt/keyrings/docker.asc
+  if cmd_missing docker; then
+    # Add Docker's official GPG key:
+    sudo apt-get update
+    sudo apt-get install ca-certificates curl
+    sudo install -m 0755 -d /etc/apt/keyrings
+    sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
+    sudo chmod a+r /etc/apt/keyrings/docker.asc
 
-# Add the repository to Apt sources:
-echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
-  $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
-  sudo tee /etc/apt/sources.list.d/docker.list > /dev/null sudo apt-get update
+    # Add the repository to Apt sources:
+    echo \
+      "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
+      $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
+      sudo tee /etc/apt/sources.list.d/docker.list > /dev/null sudo apt-get update
   
-  sudo apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin -y
+    sudo apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin -y
   
-  sudo groupadd docker
-  sudo usermod -aG docker $USER
-  newgrp docker
+    sudo groupadd docker
+    sudo usermod -aG docker $USER
+    newgrp docker
   fi
 }
 
 setup_vscode() {
-  if ! command -v code; then
-  wget "https://go.microsoft.com/fwlink/?LinkID=760868" -O /tmp/vscode.deb &&
-  sudo apt install -y /tmp/vscode.deb
+  if cmd_missing code; then
+    wget "https://go.microsoft.com/fwlink/?LinkID=760868" -O /tmp/vscode.deb &&
+    sudo apt install -y /tmp/vscode.deb
   fi
 }
 
@@ -155,6 +161,7 @@ setup_neovim() {
   # Nvchad 
   sudo apt install -y ripgrep
   git clone https://github.com/NvChad/NvChad ~/.config/nvim --depth 1
+  # TODO: Install copilot
 }
 
 setup_dev_stuff() {
@@ -171,7 +178,7 @@ setup_dev_stuff() {
 }
 
 setup_grub_customizer() {
-  if ! command -v grub-customizer; then
+  if cmd_missing grub-customizer; then
     sudo add-apt-repository ppa:danielrichter2007/grub-customizer -y
     sudo apt-get update
     sudo apt-get install grub-customizer -y
@@ -179,23 +186,25 @@ setup_grub_customizer() {
 }
 
 setup_discord() {
-  if ! command -v discord; then
-  wget "https://discord.com/api/download?platform=linux&format=deb" -O /tmp/disc
-ord.deb &&
-	sudo apt install -y /tmp/discord.deb
+  if cmd_missing discord; then
+    wget "https://discord.com/api/download?platform=linux&format=deb" -O /tmp/discord.deb
+	  sudo apt install -y /tmp/discord.deb
   fi
 }
 
 
 setup_user_apps() {
   setup_discord
+  
   # Signal
-  wget -O- https://updates.signal.org/desktop/apt/keys.asc | gpg --dearmor > signal-desktop-keyring.gpg
-cat signal-desktop-keyring.gpg | sudo tee -a /usr/share/keyrings/signal-desktop-keyring.gpg > /dev/null &&
-	echo 'deb [arch=amd64 signed-by=/usr/share/keyrings/signal-desktop-keyring.gpg] https://updates.signal.org/desktop/apt xenial main' |\
-  sudo tee -a /etc/apt/sources.list.d/signal-xenial.list &&
-  sudo apt update && sudo apt install signal-desktop &&
-  rm signal-desktop-keyring.gpg
+  if cmd_missing signal-desktop; then
+    wget -O- https://updates.signal.org/desktop/apt/keys.asc | gpg --dearmor > signal-desktop-keyring.gpg
+    cat signal-desktop-keyring.gpg | sudo tee -a /usr/share/keyrings/signal-desktop-keyring.gpg > /dev/null &&
+	  echo 'deb [arch=amd64 signed-by=/usr/share/keyrings/signal-desktop-keyring.gpg] https://updates.signal.org/desktop/apt xenial main' |\
+    sudo tee -a /etc/apt/sources.list.d/signal-xenial.list &&
+    sudo apt update && sudo apt install signal-desktop &&
+    rm signal-desktop-keyring.gpg
+  fi
 }
 
 setup_laptop() {
@@ -204,8 +213,8 @@ setup_laptop() {
 }
 
 setup_jenv(){
-  if ! command -v jenv; then
-  brew install jenv &&
+  if cmd_missing jenv; then
+    brew install jenv &&
     fish -c "fish_add_path $HOME/.jenv/bin"
   fi
 }
